+++ lib/CustomGateways/TelephoneCall/TelephoneCallGatewayFactoryBuilder.php
@@ -1,33 +1,24 @@
 <?php namespace Vankosoft\PaymentBundle\CustomGateways\TelephoneCall;
 
-use Symfony\Component\Form\FormFactoryInterface;
-use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\DependencyInjection\ContainerAwareInterface;
+use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 use Payum\Core\Bridge\Symfony\Builder\GatewayFactoryBuilder;
+use Payum\Core\GatewayFactoryInterface;
 
-class TelephoneCallGatewayFactoryBuilder extends GatewayFactoryBuilder
+class TelephoneCallGatewayFactoryBuilder extends GatewayFactoryBuilder implements ContainerAwareInterface
 {
-    /** @var FormFactoryInterface */
-    private $formFactory;
-    
-    /** @var RequestStack */
-    private $requestStack;
-    
-    public function __construct(
-        $gatewayFactoryClass,
-        FormFactoryInterface $formFactory,
-        RequestStack $requestStack
-    ) {
-        parent::__construct( $gatewayFactoryClass );
-        
-        $this->formFactory  = $formFactory;
-        $this->requestStack = $requestStack;
-    }
-    
+    use ContainerAwareTrait;
+
+    /**
+     * 
+     * {@inheritDoc}
+     * @see \Payum\Core\Bridge\Symfony\Builder\GatewayFactoryBuilder::__invoke()
+     */
     public function __invoke()
     {
-        $factory    = call_user_func_array( [$this, 'build'], func_get_args() );
-        $factory->setDependencies( $this->formFactory, $this->requestStack );
+        $gatewayFactory = call_user_func_array([$this, 'build'], func_get_args());
+        $gatewayFactory->setContainer( $this->container );
         
-        return $factory;
+        return $gatewayFactory;
     }
 }